pyproject.toml
[tool.poetry]
name = "grimoirebots"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Sergio Merino Hernández <merinhunter94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1"
gunicorn = "^20.1"
djangorestframework = "^3.14"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
